Refactor import statements: rename image_ai.py to image_ai_window.py and update references
