+++ page1.py
@@ -7,7 +7,7 @@ from PIL import Image, ImageTk
 from draggable_treeview import DraggableTreeview
 from ai_stuff import open_if_window, open_checkpoint_window, open_wait_window
 from search_pattern_window import open_pattern_window
-from image_ai import open_image_ai_window
+from image_ai_window import open_image_ai_window
 import base64
 import io
 from PIL import Image, ImageTk
